refactor(interactionTokens): move out credential offer types
